✨  Add command abstraction
