+++ symupy/abstractions/__init__.py
@@ -1,0 +1,9 @@
+"""
+Abstractions
+============
+This package contains implementations related to generic abstractions regularly implemented via metaclasses in order to provide generic software pattern behaviors.
+"""
+
+from .observer import AbsObserver
+from .subject import AbsSubject
+from .command import AbsCommand
