--- symupy/api/commands/__init__.py
@@ -1,1 +1,0 @@
-from .nocommand import NoCommand
